fix: update component refs
