+++ packages/react/package.json
@@ -50,6 +50,7 @@
     "@react-aria/tooltip": "^3.2.1",
     "@react-aria/utils": "^3.13.1",
     "@react-aria/visually-hidden": "^3.3.1",
+    "@react-spectrum/utils": "^3.7.2",
     "@react-stately/calendar": "^3.0.0",
     "@react-stately/collections": "^3.4.1",
     "@react-stately/combobox": "^3.1.1",
