+++ packages/react/src/components/Button/Button.tsx
@@ -1,12 +1,13 @@
 import type { ButtonSize, ButtonVariant } from './types';
 import type { DOMProps, StyleProps } from '../../types';
+import type { FocusableRef, PressEvent } from '@react-types/shared';
 import type { AriaButtonProps } from '@react-types/button';
-import type { PressEvent } from '@react-types/shared';
 import * as React from 'react';
 import { mergeProps, mergeRefs } from '@react-aria/utils';
 import { cx } from '../../styles';
 import { useButton } from '@react-aria/button';
 import { useButtonGroup } from '../ButtonGroup';
+import { useFocusableRef } from '@react-spectrum/utils';
 import { useFocusRing } from '@react-aria/focus';
 import { useHover } from '@react-aria/interactions';
 import { useStyles } from './Button.styles';
@@ -48,94 +49,98 @@ interface ButtonProps extends AriaButtonProps, DOMProps, StyleProps {
   onClick?(event: React.MouseEvent<HTMLButtonElement>): void;
 }
 
-const Button = React.forwardRef<ButtonElement, ButtonProps>((props, forwardedRef) => {
-  const group = useButtonGroup();
+const Button = React.forwardRef(
+  (props: ButtonProps, forwardedRef: FocusableRef<HTMLButtonElement>) => {
+    const group = useButtonGroup();
 
-  const {
-    autoFocus,
-    children,
-    className: classNameProp,
-    css,
-    isDisabled = group?.isDisabled,
-    endIcon,
-    onClick,
-    onPress,
-    size = group?.size ?? 'medium',
-    startIcon,
-    variant = group?.variant ?? 'primary',
-    ...other
-  } = props;
+    const {
+      autoFocus,
+      children,
+      className: classNameProp,
+      css,
+      isDisabled = group?.isDisabled,
+      endIcon,
+      onClick,
+      onPress,
+      size = group?.size ?? 'medium',
+      startIcon,
+      variant = group?.variant ?? 'primary',
+      ...other
+    } = props;
 
-  const buttonRef = React.useRef<HTMLButtonElement>(null);
+    const buttonRef = useFocusableRef(forwardedRef);
 
-  const handleClick = React.useCallback(
-    (event: React.MouseEvent<HTMLButtonElement>) => {
-      onClick?.(event);
-    },
-    [onClick],
-  );
+    const handleClick = React.useCallback(
+      (event: React.MouseEvent<HTMLButtonElement>) => {
+        onClick?.(event);
+      },
+      [onClick],
+    );
 
-  const handlePress = React.useCallback(
-    (event: PressEvent) => {
-      if (event.pointerType === 'keyboard' || event.pointerType === 'virtual') {
-        onClick?.(event as unknown as React.MouseEvent<HTMLButtonElement>);
-      }
+    const handlePress = React.useCallback(
+      (event: PressEvent) => {
+        if (event.pointerType === 'keyboard' || event.pointerType === 'virtual') {
+          onClick?.(event as unknown as React.MouseEvent<HTMLButtonElement>);
+        }
 
-      onPress?.(event);
-    },
-    [onClick, onPress],
-  );
+        onPress?.(event);
+      },
+      [onClick, onPress],
+    );
 
-  const { buttonProps, isPressed } = useButton(
-    {
-      ...other,
-      elementType: 'button',
+    const { buttonProps, isPressed } = useButton(
+      {
+        ...other,
+        elementType: 'button',
+        isDisabled,
+        onClick: handleClick,
+        onPress: handlePress,
+      } as AriaButtonProps,
+      buttonRef,
+    );
+    const { isFocusVisible, focusProps } = useFocusRing({ autoFocus });
+    const { hoverProps, isHovered } = useHover({ isDisabled });
+
+    const { className } = useStyles({
+      css,
+      hasEndIcon: !!endIcon,
+      hasStartIcon: !!startIcon,
       isDisabled,
-      onClick: handleClick,
-      onPress: handlePress,
-    } as AriaButtonProps,
-    buttonRef,
-  );
-  const { isFocusVisible, focusProps } = useFocusRing({ autoFocus });
-  const { hoverProps, isHovered } = useHover({ isDisabled });
+      isFocusVisible,
+      isHovered,
+      isPressed,
+      size,
+      variant,
+    });
 
-  const { className } = useStyles({
-    css,
-    hasEndIcon: !!endIcon,
-    hasStartIcon: !!startIcon,
-    isDisabled,
-    isFocusVisible,
-    isHovered,
-    isPressed,
-    size,
-    variant,
-  });
+    const classnames = cx(className, classNameProp, {
+      'manifest-button': true,
+      'manifest-button--disabled': isDisabled,
+      [`manifest-button--${size}`]: size,
+      [`manifest-button--${variant}`]: variant,
+    });
 
-  const classnames = cx(className, classNameProp, {
-    'manifest-button': true,
-    'manifest-button--disabled': isDisabled,
-    [`manifest-button--${size}`]: size,
-    [`manifest-button--${variant}`]: variant,
-  });
-
-  return (
-    <button
-      {...mergeProps(buttonProps, focusProps, hoverProps)}
-      className={classnames}
-      ref={mergeRefs(buttonRef, forwardedRef)}
-    >
-      {startIcon && (
-        <span className={cx('manifest-button__icon', 'manifest-button__icon--start')}>
-          {startIcon}
-        </span>
-      )}
-      {children}
-      {endIcon && (
-        <span className={cx('manifest-button__icon', 'manifest-button__icon--end')}>{endIcon}</span>
-      )}
-    </button>
-  );
-});
+    return (
+      <button
+        {...mergeProps(buttonProps, focusProps, hoverProps)}
+        className={classnames}
+        ref={buttonRef}
+      >
+        {startIcon && (
+          <span className={cx('manifest-button__icon', 'manifest-button__icon--start')}>
+            {startIcon}
+          </span>
+        )}
+        {children}
+        {endIcon && (
+          <span className={cx('manifest-button__icon', 'manifest-button__icon--end')}>
+            {endIcon}
+          </span>
+        )}
+      </button>
+    );
+  },
+);
 
 if (__DEV__) {
   Button.displayName = 'ManifestButton';
