+++ packages/react/src/components/Dropdown/Dropdown.context.tsx
@@ -1,7 +1,11 @@
 import * as React from 'react';
-import { useDropdown } from './useDropdown';
 
-export type DropdownContext = ReturnType<typeof useDropdown>;
+export interface DropdownContext {
+  closeOnSelect?: boolean;
+  menuProps: React.HTMLAttributes<HTMLUListElement>;
+  menuRef: React.RefObject<HTMLUListElement>;
+  onClose?(): void;
+}
 
 export const DropdownContext = React.createContext<DropdownContext | null>(null);
