--- packages/react/src/components/DropdownTrigger/DropdownTrigger.tsx
@@ -1,21 +1,0 @@
-import * as React from 'react';
-import { DropdownContext, useDropdownContext } from '../Dropdown';
-import { PopoverTrigger } from '../PopoverTrigger';
-
-type DropdownTriggerElement = React.ElementRef<'button'>;
-
-type DropdownTriggerProps = React.PropsWithChildren<unknown>;
-
-const DropdownTrigger = React.forwardRef<DropdownTriggerElement, DropdownTriggerProps>(
-  (props, forwardedRef) => {
-    const { children, ...other } = props;
-
-    const { getMenuTriggerProps } = useDropdownContext() as DropdownContext;
-
-    return (
-      <PopoverTrigger {...getMenuTriggerProps(other, forwardedRef)}>{children}</PopoverTrigger>
-    );
-  },
-);
-
-export { DropdownTrigger };
