--- packages/react/src/components/Popover/Popover.context.tsx
@@ -1,8 +1,0 @@
-import * as React from 'react';
-import { usePopover } from './usePopover';
-
-export type PopoverContext = ReturnType<typeof usePopover>;
-
-export const PopoverContext = React.createContext<PopoverContext | null>(null);
-
-export const usePopoverContext = () => React.useContext(PopoverContext);
