+++ packages/react/src/components/Popover/index.ts
@@ -1,3 +1,1 @@
 export * from './Popover';
-export * from './Popover.context';
-export * from './usePopover';
