--- packages/react/src/components/PopoverTrigger/PopoverTrigger.tsx
@@ -1,24 +1,0 @@
-import * as React from 'react';
-import { PopoverContext, usePopoverContext } from '../Popover';
-import { mergeProps } from '@react-aria/utils';
-import { Slot } from '@radix-ui/react-slot';
-
-type PopoverTriggerElement = React.ElementRef<'button'>;
-
-type PopoverTriggerProps = React.PropsWithChildren<unknown>;
-
-const PopoverTrigger = React.forwardRef<PopoverTriggerElement, PopoverTriggerProps>(
-  (props, forwardedRef) => {
-    const { children, ...other } = props;
-
-    const { getTriggerProps, state } = usePopoverContext() as PopoverContext;
-
-    return (
-      <Slot {...getTriggerProps(mergeProps(other, { onPress: () => state.open() }), forwardedRef)}>
-        {children}
-      </Slot>
-    );
-  },
-);
-
-export { PopoverTrigger };
