--- packages/react/src/components/PopoverTrigger/index.ts
@@ -1,1 +1,0 @@
-export * from './PopoverTrigger';
